requirements.txt
# Core data processing
numpy<2
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web scraping and text processing
beautifulsoup4>=4.10.0
langdetect>=1.0.9

# Machine learning
scikit-learn>=1.0.0
torch==2.8.0
torchvision==0.23.0
transformers>=4.25.0
xgboost>=1.7.0

# Image processing
pillow>=9.0.0
opencv-python-headless<4.9
imagehash>=4.3.1

# Progress bar and utilities
tqdm>=4.62.0

# Web application interface
streamlit>=1.8.0

# Natural language processing
nltk>=3.8.0
wordcloud>=1.9.2

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Mermaid for diagram rendering in Streamlit
streamlit-mermaid

# Additional utilities
sentencepiece>=0.2.0